backend/requirements.txt
# Django Core
Django==5.1.3
djangorestframework==3.15.2
django-cors-headers==4.6.0
django-filter==24.3

# Database
psycopg2-binary==2.9.10
dj-database-url==2.3.0

# Authentication
djangorestframework-simplejwt==5.4.0
django-allauth==65.3.0

# Celery & Redis
celery==5.4.0
django-celery-beat==2.7.0
django-celery-results==2.5.1
redis==5.2.1
django-redis==5.4.0

# API Documentation
drf-spectacular==0.28.0

# Image handling
Pillow==11.0.0

# Environment
python-dotenv==1.0.1
python-decouple==3.8

# Production
gunicorn==23.0.0
whitenoise==6.8.2

# Utilities
django-extensions==3.2.3
django-storages==1.14.4

# Testing
pytest==8.3.4
pytest-django==4.9.0
factory-boy==3.3.1

# Stripe payments
stripe==11.3.0
